@@ -1,5 +1,5 @@
 name: pub_semver
-version: 1.4.4
+version: 1.5.0-nullsafety
 
 description: >-
  Versions and version constraints implementing pub's versioning policy. This
@@ -7,10 +7,64 @@ description: >-
 homepage: https://github.com/dart-lang/pub_semver
 
 environment:
- sdk: '>=2.0.0 <3.0.0'
+ sdk: '>=2.10.0-0 <2.10.0'
 
 dependencies:
   collection: ^1.0.0
 
 dev_dependencies:
   test: ^1.0.0
+
+dependency_overrides:
+ async:
+  git: git://github.com/dart-lang/async.git
+ boolean_selector:
+  git: git://github.com/dart-lang/boolean_selector.git
+ charcode:
+  git: git://github.com/dart-lang/charcode.git
+ collection:
+  git: git://github.com/dart-lang/collection.git
+ js:
+  git:
+   url: git://github.com/dart-lang/sdk.git
+   path: pkg/js
+   ref: 2-10-pkgs
+ matcher:
+  git: git://github.com/dart-lang/matcher.git
+ meta:
+  git:
+   url: git://github.com/dart-lang/sdk.git
+   path: pkg/meta
+   ref: 2-10-pkgs
+ pedantic:
+  git: git://github.com/dart-lang/pedantic.git
+ pool:
+  git: git://github.com/dart-lang/pool.git
+ source_maps:
+  git: git://github.com/dart-lang/source_maps.git
+ source_map_stack_trace:
+  git: git://github.com/dart-lang/source_map_stack_trace.git
+ source_span:
+  git: git://github.com/dart-lang/source_span.git
+ stack_trace:
+  git: git://github.com/dart-lang/stack_trace.git
+ stream_channel:
+  git: git://github.com/dart-lang/stream_channel.git
+ string_scanner:
+  git: git://github.com/dart-lang/string_scanner.git
+ term_glyph:
+  git: git://github.com/dart-lang/term_glyph.git
+ test_api:
+  git:
+   url: git://github.com/dart-lang/test.git
+   path: pkgs/test_api
+ test_core:
+  git:
+   url: git://github.com/dart-lang/test.git
+   path: pkgs/test_core
+ test:
+  git:
+   url: git://github.com/dart-lang/test.git
+   path: pkgs/test
+ typed_data:
+  git: git://github.com/dart-lang/typed_data.git
