@@ -19,7 +19,7 @@ bool allowsLower(VersionRange range1, VersionRange range2) {
   if (range1.min == null) return range2.min != null;
   if (range2.min == null) return false;
 
-  var comparison = range1.min.compareTo(range2.min);
+  var comparison = range1.min!.compareTo(range2.min!);
   if (comparison == -1) return true;
   if (comparison == 1) return false;
   return range1.includeMin && !range2.includeMin;
@@ -30,7 +30,7 @@ bool allowsHigher(VersionRange range1, VersionRange range2) {
   if (range1.max == null) return range2.max != null;
   if (range2.max == null) return false;
 
-  var comparison = range1.max.compareTo(range2.max);
+  var comparison = range1.max!.compareTo(range2.max!);
   if (comparison == 1) return true;
   if (comparison == -1) return false;
   return range1.includeMax && !range2.includeMax;
@@ -41,7 +41,7 @@ bool allowsHigher(VersionRange range1, VersionRange range2) {
 bool strictlyLower(VersionRange range1, VersionRange range2) {
   if (range1.max == null || range2.min == null) return false;
 
-  var comparison = range1.max.compareTo(range2.min);
+  var comparison = range1.max!.compareTo(range2.min!);
   if (comparison == -1) return true;
   if (comparison == 1) return false;
   return !range1.includeMax || !range2.includeMin;
